requirements.txt
fastapi[all]
sqlalchemy
asyncpg
python-dotenv
boto3
